UnmanagedThunks, InternalCalls, WorldSpace
- Removed Event Status
- Removed EventDispacher::DispachEventAlways
- Removed TransformUtils
- Removed WorldSpaceTransform functions from Entity
- Added GetWorldSpaceTransform to Scene

-Added Wrapper for UnmanagedThunks
- Switch from MonoMethod to UnmanagedThunks for most Methods in Entity
- Added mono string utils

- Chanched syntax of moste InternalCalls
- Removed TimeStep form ScriptCore
- Removed Events from ScriptCore
- Renamed Editor.UI to EditorUI
- Chanched Entity.ID to field instead of property
- Added Phyiscs2D functions to ScriptCore

+++ Shark/src/Platform/Windows/WindowsUtils.cpp
@@ -198,7 +198,7 @@ namespace Shark {
 		std::string envVar;
 		envVar.resize(bufferSize);
 
-		DWORD result = GetEnvironmentVariableA(name.c_str(), envVar.data(), envVar.size());
+		DWORD result = GetEnvironmentVariableA(name.c_str(), envVar.data(), (DWORD)envVar.size());
 		if (result == 0)
 		{
 			DWORD errorCode = GetLastError();
